src/main/java/problems/p0000_0100/p0041_first_missing_positive/Solution.java
package problems.p0000_0100.p0041_first_missing_positive;

class Solution {
    public int firstMissingPositive(int[] nums) {
        int i = 0;
        int n = nums.length;
        while (i < n) {
            if (nums[i] > 0 && nums[i] <= n && nums[i] != nums[nums[i] - 1]) {
                swap(nums, i, nums[i] - 1);
            } else {
                i++;
            }
        }
        i = -1;
        while (++i < n) {
            if (nums[i] != i + 1) {
                return i + 1;
            }
        }
        return nums.length + 1;
    }

    private void swap(int[] nums, int i, int j) {
        final int t = nums[i];
        nums[i] = nums[j];
        nums[j] = t;
    }
}
